refactor: Introduce intermediate classes for ES/OS properties
