+++ configuration/src/main/java/io/camunda/configuration/DocumentBasedSecondaryStorageDatabase.java
@@ -1,0 +1,175 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+package io.camunda.configuration;
+
+import io.camunda.configuration.UnifiedConfigurationHelper.BackwardsCompatibilityMode;
+import java.util.Set;
+import org.springframework.boot.context.properties.NestedConfigurationProperty;
+
+public abstract class DocumentBasedSecondaryStorageDatabase
+    extends SecondaryStorageDatabase<DocumentBasedHistory> {
+
+  /** Prefix to apply to the indexes. */
+  private String indexPrefix = "";
+
+  /** Name of the cluster */
+  private String clusterName = databaseName().toLowerCase();
+
+  /** How many shards Elasticsearch uses for all Tasklist indices. */
+  private int numberOfShards = 1;
+
+  @NestedConfigurationProperty private Security security = new Security(databaseName());
+
+  @NestedConfigurationProperty
+  private DocumentBasedHistory history = new DocumentBasedHistory(databaseName());
+
+  @Override
+  public String getUrl() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".url",
+        super.getUrl(),
+        String.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        legacyUrlProperties());
+  }
+
+  @Override
+  public String getUsername() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".username",
+        super.getUsername(),
+        String.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        legacyUsernameProperties());
+  }
+
+  @Override
+  public String getPassword() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".password",
+        super.getPassword(),
+        String.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        legacyPasswordProperties());
+  }
+
+  @Override
+  public DocumentBasedHistory getHistory() {
+    return history;
+  }
+
+  @Override
+  public void setHistory(final DocumentBasedHistory history) {
+    this.history = history;
+  }
+
+  @Override
+  protected abstract String databaseName();
+
+  public Security getSecurity() {
+    return security;
+  }
+
+  public void setSecurity(final Security security) {
+    this.security = security;
+  }
+
+  public String getClusterName() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".cluster-name",
+        clusterName,
+        String.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        legacyClusterNameProperties());
+  }
+
+  public void setClusterName(final String clusterName) {
+    this.clusterName = clusterName;
+  }
+
+  public String getIndexPrefix() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".index-prefix",
+        indexPrefix,
+        String.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        indexPrefixLegacyProperties());
+  }
+
+  public void setIndexPrefix(final String indexPrefix) {
+    this.indexPrefix = indexPrefix;
+  }
+
+  public int getNumberOfShards() {
+    return UnifiedConfigurationHelper.validateLegacyConfiguration(
+        prefix() + ".number-of-shards",
+        numberOfShards,
+        Integer.class,
+        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
+        legacyNumberOfShardsProperties());
+  }
+
+  public void setNumberOfShards(final int numberOfShards) {
+    this.numberOfShards = numberOfShards;
+  }
+
+  private String prefix() {
+    return "camunda.data.secondary-storage." + databaseName().toLowerCase();
+  }
+
+  private Set<String> legacyUrlProperties() {
+    final String dbName = databaseName().toLowerCase();
+    return Set.of(
+        "camunda.database.url",
+        "camunda.operate." + dbName + ".url",
+        "camunda.tasklist." + dbName + ".url",
+        "zeebe.broker.exporters.camundaexporter.args.connect.url");
+  }
+
+  private Set<String> legacyClusterNameProperties() {
+    final String dbName = databaseName().toLowerCase();
+    return Set.of(
+        "camunda.database.clusterName",
+        "camunda.operate." + dbName + ".clusterName",
+        "camunda.tasklist." + dbName + ".clusterName",
+        "zeebe.broker.exporters.camundaexporter.args.connect.clusterName");
+  }
+
+  private Set<String> legacyUsernameProperties() {
+    final String dbName = databaseName().toLowerCase();
+    return Set.of(
+        "camunda.database.username",
+        "camunda.operate." + dbName + ".username",
+        "camunda.tasklist." + dbName + ".username",
+        "zeebe.broker.exporters.camundaexporter.args.connect.username");
+  }
+
+  private Set<String> legacyPasswordProperties() {
+    final String dbName = databaseName().toLowerCase();
+    return Set.of(
+        "camunda.database.password",
+        "camunda.operate." + dbName + ".password",
+        "camunda.tasklist." + dbName + ".password",
+        "zeebe.broker.exporters.camundaexporter.args.connect.password");
+  }
+
+  private Set<String> indexPrefixLegacyProperties() {
+    final String dbName = databaseName().toLowerCase();
+    return Set.of(
+        "camunda.database.indexPrefix",
+        "camunda.tasklist." + dbName + ".indexPrefix",
+        "camunda.operate." + dbName + ".indexPrefix",
+        "zeebe.broker.exporters.camundaexporter.args.index.indexPrefix");
+  }
+
+  private Set<String> legacyNumberOfShardsProperties() {
+    return Set.of(
+        "camunda.database.index.numberOfShards",
+        "zeebe.broker.exporters.camundaexporter.args.index.numberOfShards");
+  }
+}
