+++ configuration/src/main/java/io/camunda/configuration/SecondaryStorageDatabase.java
@@ -7,21 +7,18 @@
  */
 package io.camunda.configuration;
 
-import io.camunda.configuration.UnifiedConfigurationHelper.BackwardsCompatibilityMode;
-import java.util.Set;
-import org.springframework.boot.context.properties.NestedConfigurationProperty;
-
-public abstract class SecondaryStorageDatabase {
+/**
+ * Base class for configuration of databases used as secondary storage.
+ *
+ * <p>Note: For now there only is support for document-based databases, but in the future also
+ * relational databases will be supported.
+ *
+ * @param <T> the type of history configuration this database supports.
+ */
+public abstract class SecondaryStorageDatabase<T> {
 
   /** Endpoint for the database configured as secondary storage. */
   private String url = "http://localhost:9200";
-
-  /** Name of the cluster */
-  private String clusterName = databaseName().toLowerCase();
-
-  @NestedConfigurationProperty private Security security = new Security(databaseName());
-
-  @NestedConfigurationProperty private History history = new History(databaseName());
 
   /** Username for the database configured as secondary storage. */
   private String username = "";
@@ -29,19 +26,8 @@ public abstract class SecondaryStorageDatabase {
   /** Password for the database configured as secondary storage. */
   private String password = "";
 
-  /** Prefix to apply to the indexes. */
-  private String indexPrefix = "";
-
-  /** How many shards Elasticsearch uses for all Tasklist indices. */
-  private int numberOfShards = 1;
-
   public String getUrl() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".url",
-        url,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        legacyUrlProperties());
+    return url;
   }
 
   public void setUrl(final String url) {
@@ -49,12 +35,7 @@ public abstract class SecondaryStorageDatabase {
   }
 
   public String getUsername() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".username",
-        username,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        legacyUsernameProperties());
+    return username;
   }
 
   public void setUsername(final String username) {
@@ -62,126 +43,19 @@ public abstract class SecondaryStorageDatabase {
   }
 
   public String getPassword() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".password",
-        password,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        legacyPasswordProperties());
+    return password;
   }
 
   public void setPassword(final String password) {
     this.password = password;
   }
 
-  public Security getSecurity() {
-    return security;
-  }
+  public abstract T getHistory();
 
-  public void setSecurity(final Security security) {
-    this.security = security;
-  }
-
-  public String getClusterName() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".cluster-name",
-        clusterName,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        legacyClusterNameProperties());
-  }
-
-  public void setClusterName(final String clusterName) {
-    this.clusterName = clusterName;
-  }
-
-  public String getIndexPrefix() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".index-prefix",
-        indexPrefix,
-        String.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        indexPrefixLegacyProperties());
-  }
-
-  public void setIndexPrefix(final String indexPrefix) {
-    this.indexPrefix = indexPrefix;
-  }
-
-  public int getNumberOfShards() {
-    return UnifiedConfigurationHelper.validateLegacyConfiguration(
-        prefix() + ".number-of-shards",
-        numberOfShards,
-        Integer.class,
-        BackwardsCompatibilityMode.SUPPORTED_ONLY_IF_VALUES_MATCH,
-        legacyNumberOfShardsProperties());
-  }
-
-  public void setNumberOfShards(final int numberOfShards) {
-    this.numberOfShards = numberOfShards;
-  }
-
-  public History getHistory() {
-    return history;
-  }
-
-  public void setHistory(final History history) {
-    this.history = history;
-  }
+  public abstract void setHistory(final T history);
 
   private String prefix() {
     return "camunda.data.secondary-storage." + databaseName().toLowerCase();
-  }
-
-  private Set<String> legacyUrlProperties() {
-    final String dbName = databaseName().toLowerCase();
-    return Set.of(
-        "camunda.database.url",
-        "camunda.operate." + dbName + ".url",
-        "camunda.tasklist." + dbName + ".url",
-        "zeebe.broker.exporters.camundaexporter.args.connect.url");
-  }
-
-  private Set<String> legacyClusterNameProperties() {
-    final String dbName = databaseName().toLowerCase();
-    return Set.of(
-        "camunda.database.clusterName",
-        "camunda.operate." + dbName + ".clusterName",
-        "camunda.tasklist." + dbName + ".clusterName",
-        "zeebe.broker.exporters.camundaexporter.args.connect.clusterName");
-  }
-
-  private Set<String> legacyUsernameProperties() {
-    final String dbName = databaseName().toLowerCase();
-    return Set.of(
-        "camunda.database.username",
-        "camunda.operate." + dbName + ".username",
-        "camunda.tasklist." + dbName + ".username",
-        "zeebe.broker.exporters.camundaexporter.args.connect.username");
-  }
-
-  private Set<String> legacyPasswordProperties() {
-    final String dbName = databaseName().toLowerCase();
-    return Set.of(
-        "camunda.database.password",
-        "camunda.operate." + dbName + ".password",
-        "camunda.tasklist." + dbName + ".password",
-        "zeebe.broker.exporters.camundaexporter.args.connect.password");
-  }
-
-  private Set<String> indexPrefixLegacyProperties() {
-    final String dbName = databaseName().toLowerCase();
-    return Set.of(
-        "camunda.database.indexPrefix",
-        "camunda.tasklist." + dbName + ".indexPrefix",
-        "camunda.operate." + dbName + ".indexPrefix",
-        "zeebe.broker.exporters.camundaexporter.args.index.indexPrefix");
-  }
-
-  private Set<String> legacyNumberOfShardsProperties() {
-    return Set.of(
-        "camunda.database.index.numberOfShards",
-        "zeebe.broker.exporters.camundaexporter.args.index.numberOfShards");
   }
 
   protected abstract String databaseName();
